requirements.txt
# Alpaca API for trading & market data
alpaca-py

# Data handling & numerical analysis
pandas
numpy

# Time and date utilities
python-dateutil

# Backtesting framework
backtrader

# Environment variable management
python-dotenv

# HTTP requests for Telegram notifications
requests
